backend/requirements.txt
absl-py==2.3.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.40.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==3.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
Pillow==10.0.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1
